trans removeAll.js to reset.ts
